requirements.txt
Django==4.1
python-magic==0.4.27
requests==2.28.1
google-api-python-client==2.58.0
asyncio==3.4.3
youtube-dl
